fix topic repository

+++ base-data/src/main/java/com/example/android_data/Topic.kt
@@ -6,7 +6,7 @@ import androidx.room.PrimaryKey
 
 @Entity
 data class Topic(
-    @PrimaryKey(autoGenerate = true) val id: Long = 0,
+    @PrimaryKey(autoGenerate = true) val id: Int = 0,
     @ColumnInfo(name = "title") val title: String? = null,
     @ColumnInfo(name = "content") val content: String? = null,
     @ColumnInfo(name = "page") val page: Int = 0
